Update intro.md
This change highlights the built-in functions default behavior when something wrong happens during its execution. In exception or error scenarios, the runtime behavior is to prevent the exception from escaping and instead return the default value of the output data type.

This has always been the case in all OutSystems products.